addressed review comments, mostly moving repeated code to the BaseData class
