+++ .gitignore
@@ -1,13 +1,13 @@
-.vscode
+# Python virtualenv
 .venv
+
+# VS Code specifics
+.vscode
 .devcontainer
-wrfout_d02_2019-03-05_12_00_00.nc
-tests/Lakes/input/hrrr.20190305.full
 
 # Examples files
-examples/*/*/maxus*.nc
-*/*/output/RME
-*/*/log.*
+wrfout_d02_2019-03-05_12_00_00.nc
+tests/Lakes/input/hrrr.20190305.full
 
 # Byte-compiled / optimized / DLL files
 __pycache__/
@@ -76,35 +76,7 @@ target/
 # git version stuff:
 katana/gitinfo.py
 
-#Auto generate config files
-*config_full.ini
-test_data/config.ini
-
-#Auto backup
-*/input_backup/*
-
 #ignore autogenerated config doc
 docs/auto_config.rst
 docs/modules.rst
 docs/smrf.rst
-
-#Ignore results in test_data
-test_data/output/*
-test_data/tmp/*
-
-#Auto generated config.
-test_data/config.ini
-
-#ignore maxus.
-test_data/topo/maxus_100window.nc
-test_data/topo/maxus.nc
-examples/* *.nc
-examples/reynolds_mountain_east/output/*
-examples/simple_hill/* *.csv
-*/input_backup/*
-test_data/output/*
-*/tmp/*
-
-tests/RME/output/*
-tests/RME/topo/*.asc
-tests/RME/topo/*.prj
